+++ GIAN07/player/loadout/homing_loadout.h
@@ -10,17 +10,17 @@ class HomingLoadout final : public PlayerLoadout {
 public:
   HomingLoadout();
 
-  void FireMain(Player &player, uint8_t tier, bool focused) override;
-  void FireSub(Player &player, uint8_t tier, bool focused) override;
+  void FireMain(Player &player, int tier, bool focused) override;
+  void FireSub(Player &player, int tier, bool focused) override;
   void UpdateBomb(Player &player, EnemyManager &enemies, EffectManager &effects,
-                  uint16_t remaining) override;
+                  int remaining) override;
   void Reset() override { shot_phase_ = 0; }
 
 private:
-  uint8_t shot_phase_ = 0;
+  int shot_phase_ = 0;
 
-  void FireMainNormal(Player &player, uint8_t tier);
-  static void FireMainFocused(Player &player, uint8_t tier);
-  void FireSubNormal(Player &player, uint8_t tier);
-  void FireSubFocused(Player &player, uint8_t tier);
+  void FireMainNormal(Player &player, int tier);
+  static void FireMainFocused(Player &player, int tier);
+  void FireSubNormal(Player &player, int tier);
+  void FireSubFocused(Player &player, int tier);
 };
