+++ GIAN07/player/loadout/wide_loadout.h
@@ -10,19 +10,19 @@ class WideLoadout final : public PlayerLoadout {
 public:
   WideLoadout();
 
-  void FireMain(Player &player, uint8_t tier, bool focused) override;
-  void FireSub(Player &player, uint8_t tier, bool focused) override;
+  void FireMain(Player &player, int tier, bool focused) override;
+  void FireSub(Player &player, int tier, bool focused) override;
   void UpdateBomb(Player &player, EnemyManager &enemies, EffectManager &effects,
-                  uint16_t remaining) override;
+                  int remaining) override;
   void DrawBombBackground(const Player &player,
-                          uint16_t remaining) const override;
+                          int remaining) const override;
   void Reset() override { shot_phase_ = 0; }
 
 private:
-  uint8_t shot_phase_ = 0;
+  int shot_phase_ = 0;
 
-  void FireMainNormal(Player &player, uint8_t tier);
-  static void FireMainFocused(Player &player, uint8_t tier);
-  void FireSubNormal(Player &player, uint8_t tier);
-  void FireSubFocused(Player &player, uint8_t tier);
+  void FireMainNormal(Player &player, int tier);
+  static void FireMainFocused(Player &player, int tier);
+  void FireSubNormal(Player &player, int tier);
+  void FireSubFocused(Player &player, int tier);
 };
